Implement a mixin for configurable zombie conversion.
Define a configuration type and a way for the mixin to access the configuration.

The mixin optionally supresses the vanilla behaviour when a zombie kills a villager. Instead of having the probability
of conversion to a zombie villager be dependent on difficulty, this mixin will instead apply a constant probability
regardless of difficulty. The probability (aka the "conversion rate") can be anywhere from 0% to 100%, inclusive.

When a villager is converted to a zombie, the mixin exactly replicates the vanilla behaviour. Once cured, the villager
will have the same profession and trades as before (supposing it had a profession). It will even keep is baby/adult
status, as well as any xp, gossip, etc.

The configuration is read from the `villager_conversion_rate.json` file in the standard fabric `config/` directory. If
it doesn't exist (for example, on the first run with the mod installed) it will be created with the default values. The
configuration can be edited directly, in which case MineCraft will have to be restarted. When loaded in a client, a
better way to configure the mod is using the built-in ModMenu support. Not only does this provide a nice GUI for editing
the configuration, it also allows the changes to be applied immediately. There is a slight visual difference between
editing the configuration file and using ModMenu: the configuration file uses a proper rate in the range `[0, 1]`, while
the ModMenu screen represents this value as an integer percentage from 0 to 100. In reality, the mod always uses the
rate, with the percentage being converted to a rate when the configuration is saved.